requirements.txt
fastapi==0.132.0
uvicorn[standard]==0.41.0
anthropic==0.78.0
python-dotenv==1.1.0
